src/router_worker/requirements.txt
kafka-python
requests
redis
psycopg2-binary
SQLAlchemy
pandas
